# server/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "Fiber Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/game/": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "create new game.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "create new game.",
                "parameters": [
                    {
                        "description": "request",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_IvaCheMih_chess_server_domains_game_dto.RequestedCreateGame"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/game/:gameId/history": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get history.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "get history.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gameId",
                        "name": "gameId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "userId",
                        "name": "userId",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/game/:gameId/move": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "do move.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "move"
                ],
                "summary": "do move.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gameId",
                        "name": "gameId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "userId",
                        "name": "userId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "move",
                        "name": "move",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_IvaCheMih_chess_server_domains_game_dto.RequestDoMove"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/game/{gameId}/board": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get board.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board"
                ],
                "summary": "get board.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "gameId",
                        "name": "gameId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_IvaCheMih_chess_server_domains_game_dto.GetBoardResponse"
                        }
                    }
                }
            }
        },
        "/session/": {
            "post": {
                "description": "create new session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "create new session.",
                "parameters": [
                    {
                        "description": "request",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequestUserIdAndPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/": {
            "post": {
                "description": "create new user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "create new user.",
                "parameters": [
                    {
                        "description": "request",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequestPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.RequestPassword": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.RequestUserIdAndPassword": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_IvaCheMih_chess_server_domains_game_dto.BoardCellEntity": {
            "type": "object",
            "properties": {
                "figureId": {
                    "type": "integer"
                },
                "index": {
                    "type": "integer"
                }
            }
        },
        "github_com_IvaCheMih_chess_server_domains_game_dto.GetBoardResponse": {
            "type": "object",
            "properties": {
                "boardCells": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_IvaCheMih_chess_server_domains_game_dto.BoardCellEntity"
                    }
                }
            }
        },
        "github_com_IvaCheMih_chess_server_domains_game_dto.RequestDoMove": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "github_com_IvaCheMih_chess_server_domains_game_dto.RequestedCreateGame": {
            "type": "object",
            "properties": {
                "isWhite": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "description": "JWT security accessToken. Please add it in the format \"Bearer {AccessToken}\" to authorize your requests.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}